\name{selectmcdm}
\alias{selectmcdm}
\encoding{UTF-8}
\title{Selection of Multi-Criteria Decision-Making Methods}
\description{
 This functions selects the most suitable MCDM method for a given problem by evaluating the ranking results 
 of various MCDM methods (\code{rankmat}) against the decision matrix of alternatives (\code{dmatrix}). 
 The function proposes the MCDM method that best ranks the alternatives with the highest composite criterion 
 value among the "top m" alternatives. The number of "top" alternatives (\code{mval}) can be determined manually 
 or dynamically using various methods (percentage-based, entropy-based, rule-based, or top-k frequency).
}
\usage{
selectmcdm(rankmat, dmatrix, bcvec, weights, mval = NULL,
  mvmet = "percent", mvpars = list(perc = 0.25, rule = c(2, 4, 5), k=5, r=3),
  verbose = TRUE)
}
\arguments{
  \item{rankmat}{
    A matrix or data frame of ranks generated by different MCDM methods for the alternatives. Rows should represent MCDM methods, and columns should represent alternatives. Each cell should contain the rank assigned by the respective MCDM method to that alternative (1 being the best).
  }
  \item{dmatrix}{
    A decision matrix (matrix or data frame) containing the performance values of alternatives across different criteria. Rows should represent alternatives, and columns should represent criteria.
  }
  \item{bcvec}{
    A numeric vector indicating the benefit/cost type for each criterion. \code{1} indicates a benefit criterion (to be maximized), and \code{-1} indicates a cost criterion (to be minimized). The length of the vector must be equal to the number of columns in \code{dmatrix}.
  }
  \item{weights}{
    A numeric vector specifying the weights assigned to each criterion. The length of the vector must be equal to the number of columns in \code{dmatrix}.
  }
  \item{mval}{
    The number of "top" alternatives to consider. If \code{NULL}, it will be determined dynamically based on the \code{mvmet} and \code{mvpars} arguments.
  }
  \item{mvmet}{
    A character string specifying the method to use for dynamic determination of \code{mval} when \code{mval} is \code{NULL}. Options are:
    \itemize{
      \item \code{"percent"}: Based on a specified percentage of alternatives (\code{mvpars$perc}).
      \item \code{"entropy"}: Based on the entropy of the rank distribution.
      \item \code{"rule"}: Based on predefined rules (\code{mvpars$rule}) related to the total number of alternatives.
      \item \code{"topk"}: Based on the frequency of alternatives appearing in the top \code{k} positions (\code{mvpars$k}, \code{mvpars$r}).
    }
  }
  
  \item{mvpars}{
    A list of additional parameters for the \code{mvmet} method.
    \itemize{
      \item \code{perc}: Percentage of alternatives to select. A numeric value between 0 and 1 for \code{mvmet = "percent"}.
      \item \code{rule}: A numeric vector of length 3 for \code{mvmet = "rule"} (e.g., c(2, 4, 5) for n < 5, 5 <= n <= 10, n > 10 respectively).
      \item \code{k}: Integer, specifying the number of top-ranked alternatives to consider for \code{mvmet = "topk"}.
      \item \code{r}: Integer, specifying the minimum frequency threshold for an alternative to be counted in the top-k positions for \code{mvmet = "topk"}.
    }
  }
  \item{verbose}{
    Logical; if \code{TRUE}, detailed messages about the process will be printed to the console.
  }
}
\details{
  The function evaluates multiple MCDM methods and selects the most suitable one 
  based on cumulative performance values. As the extensions to the proposed method by Bandyopadhyay (2021),
  if \code{mval} is not provided, it is determined dynamically using one of the following methods:
  \itemize{
    \item{\code{"percent"}: Selects a percentage of alternatives based on \code{mvpars$perc}.}
    \item{\code{"entropy"}: Uses Shannon entropy to determine \code{mval} dynamically.}
    \item{\code{"rule"}: Applies predefined thresholds to determine \code{mval}}.
    \item{\code{"topk"}: Selects alternatives appearing frequently in the top \code{k} positions.}
  }
}
\value{
  Returns a list with the following components:
  \item{m}{The calculated or user-specified number of top alternatives.}
  \item{suggestedmcdm}{The name of the MCDM method that yielded the highest cumulative criterion value for the given \code{m} alternatives.}
  \item{highcumval}{The highest cumulative criterion value achieved by the suggested MCDM method for the \code{m} alternatives.}
  \item{all_cumulvalm}{A named vector containing the cumulative criterion values for each MCDM method at the \code{m}-th alternative.}
}

\references{
Bandyopadhyay, S. (2021). Comparison among multi-criteria decision analysis techniques: A novel method. \emph{Progress in Artificial Intelligence, 10}(2), 195-216. <doi:10.1007/s13748-021-00235-5>
}

\author{
  Cagatay Cebeci
}

\examples{
# Decison matrix
datadf <- data.frame(
  P = c(4800, 10000, 3890, 7800, 4450, 5000, 1990, 3000, 4400, 4000),
  MSS = c(2700, 3000, 4000, 2000, 3000, 2700, 3000, 2200, 2400, 2800),
  MSL = c(12.7, 20, 13, 8, 14.5, 12.8, 40, 12.7, 14.3, 15.2),
  LC = c(50, 70, 10, 8, 6.8, 10, 6, 6.5, 14.3, 8),
  MC = c(3000, 4000, 500, 1500, 1500, 2000, 350, 1500, 500, 1000),
  MR = c(980, 1000, 800, 900, 1040, 500, 510, 1200, 850, 900),
  W = c(150, 200, 210, 100, 500, 350, 200, 230, 200, 250),
  WRT = c(6, 1, 3, 1, 5, 1, 2, 1, 3, 2)
)
rownames(datadf) <- paste0("Rob.", 1:10)
dmat <- as.matrix(datadf)

# Benefit-Cost vector
bc <- c(
  P = -1,  # Price
  MSS = 1, # Maximum sewing speed (rpm)
  MSL = 1, # Maximum stitch length (mm)
  LC = 1,  # Load capacity (kg)
  MC = 1,  # Memory capacity
  MR = 1,  # Manipulator reach
  W = -1,  # Weight (kg)
  WRT = 1  # Warranty (years)
)

# Weights
uwei <- c(
  P = 0.172414,
  MSS = 0.172414,
  MSL = 0.137931,
  LC = 0.12069,
  MC = 0.013448,
  MR = 0.137931,
  W = 0.068966,
  WRT = 0.086207
)

rankdf <- data.frame(
  R1 = c(5, 10, 10, 8, 9, 8, 3, 2, 4, 2, 2, 7, 1, 4),
  R2 = c(10, 9, 9, 10, 7, 10, 1, 1, 1, 1, 1, 10, 2, 1),
  R3 = c(3, 6, 8, 4, 1, 5, 4, 5, 8, 2, 7, 3, 3, 6),
  R4 = c(9, 3, 4, 5, 8, 9, 6, 3, 2, 6, 4, 9, 7, 3),
  R5 = c(7, 7, 6, 9, 3, 6, 2, 4, 5, 4, 3, 5, 5, 2),
  R6 = c(8, 1, 1, 7, 4, 7, 7, 6, 3, 5, 5, 8, 10, 5),
  R7 = c(1, 5, 7, 1, 2, 1, 9, 8, 10, 3, 10, 1, 4, 10),
  R8 = c(2, 2, 5, 3, 6, 2, 10, 10, 9, 7, 9, 2, 6, 9),
  R9 = c(6, 4, 2, 2, 10, 3, 8, 9, 6, 9, 8, 6, 9, 8),
  R10 = c(4, 8, 3, 6, 5, 4, 5, 7, 7, 8, 6, 4, 8, 7)
)
rownames(rankdf) <- c("TOPSIS", "MAUT", "MACBETH", 
   "ORESTE", "VIKOR", "SIR", "EVAMIX", "ARAS", "MOORA",
   "COPRAS", "WASPAS", "TODIM", "MABAC", "MARE")
rankmat <- as.matrix(rankdf)

# With a user-defined specified mval
results_user <- selectmcdm(rankmat, dmat, bc, uwei, mval = 4, verbose = FALSE)
print(results_user)

# With percentage-based mval
results_percent <- selectmcdm(rankmat, dmat, bc, uwei, 
   mvmet = "percent", mvpars = list(perc = 0.25), verbose = FALSE)
print(results_percent)

# With entropy-based mval
results_entropy <- selectmcdm(rankmat, dmat, bc, uwei, 
   mvmet = "entropy", verbose = FALSE)
print(results_entropy)

# With rule-based mval
results_rule <- selectmcdm(rankmat, dmat, bc, uwei, mvmet = "rule",
   mvpars = list(rule = c(2, 6, 8)), verbose = FALSE)
print(results_rule)

# With top-k based mval
results_topk_m <- selectmcdm(rankmat, dmat, bc, uwei, mvmet = "topk",
   mvpars = list(k = 3, r = nrow(dmat)/2), verbose = FALSE)
print(results_topk_m)
}

\keyword{MCDM}
\keyword{Multi-criteria decision making}
\keyword{ranking}
\keyword{performance evaluation}
